// apps/api/src/auth/config.interface.ts
import { AppInfo } from 'supertokens-node/types';

export const ConfigInjectionToken = 'ConfigInjectionToken';

export type AuthModuleConfig = {
    appInfo: AppInfo;
    connectionURI: string;
    apiKey: string;
    googleClientId: string;
    googleClientSecret: string;
    DashboardApiKey: string;
};